% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lt_use_conduct.R, R/lt_use_index.R
\name{lt_use_conduct}
\alias{lt_use_conduct}
\alias{lt_use_conduct}
\title{Add a code of conduct}
\usage{
lt_use_conduct(type)

lt_use_conduct(type)
}
\description{
Adds a `CODE_OF_CONDUCT.md` file to the project's top-level directory.

Adds an `INDEX.md` file to the project's top-level directory.
}
